Move App\Entity\User to \App\Entity\User\User
