+++ src/Entity/User/User.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Entity;
+namespace App\Entity\User;
 
 use App\Entity\Track\VersionRating;
 use Doctrine\Common\Collections\ArrayCollection;
